+++ src/lib/tools.ts
@@ -20,15 +20,14 @@ export const envDelimiter: string = process.platform === 'win32' ? ';' : ':';
  *
  * @warning this should match the value on the package.json otherwise the extension
  * won't work at all
- *
- * @param folder `optional` WorkspaceFolder to search
+ * @param path `optional` string for `pattern` path to include
  * @returns tuple of DocumentSelector
  */
-export function FortranDocumentSelector(folder?: vscode.WorkspaceFolder) {
-  if (folder) {
+export function FortranDocumentSelector(path?: string) {
+  if (path) {
     return [
-      { scheme: 'file', language: 'FortranFreeForm', pattern: `${folder.uri.fsPath}/**/*` },
-      { scheme: 'file', language: 'FortranFixedForm', pattern: `${folder.uri.fsPath}/**/*` },
+      { scheme: 'file', language: 'FortranFreeForm', pattern: `${path}/**/*` },
+      { scheme: 'file', language: 'FortranFixedForm', pattern: `${path}/**/*` },
     ];
   } else {
     return [
@@ -36,6 +35,54 @@ export function FortranDocumentSelector(folder?: vscode.WorkspaceFolder) {
       { scheme: 'file', language: 'FortranFixedForm' },
     ];
   }
+}
+
+export function isFortran(document: vscode.TextDocument): boolean {
+  return (
+    FortranDocumentSelector().some(e => e.scheme === document.uri.scheme) &&
+    FortranDocumentSelector().some(e => e.language === document.languageId)
+  );
+}
+
+//
+// Taken with minimal alterations from lsp-multi-server-sample
+//
+
+/**
+ * Return in ascending order the workspace folders in an array of strings
+ * @returns sorted workspace folders
+ */
+export function sortedWorkspaceFolders(): string[] | undefined {
+  const workspaceFolders = vscode.workspace.workspaceFolders
+    ? vscode.workspace.workspaceFolders
+        .map(folder => {
+          let result = folder.uri.toString();
+          if (result.charAt(result.length - 1) !== '/') result = result + '/';
+          return result;
+        })
+        .sort((a, b) => {
+          return a.length - b.length;
+        })
+    : [];
+  return workspaceFolders;
+}
+
+/**
+ * Locate the top most workspace folder for a given file
+ * @param folder workspace folder
+ * @returns outer most workspace folder
+ */
+export function getOuterMostWorkspaceFolder(
+  folder: vscode.WorkspaceFolder
+): vscode.WorkspaceFolder {
+  const sorted = sortedWorkspaceFolders();
+  for (const element of sorted) {
+    let uri = folder.uri.toString();
+    if (uri.charAt(uri.length - 1) !== '/') uri = uri + '/';
+    if (uri.startsWith(element))
+      return vscode.workspace.getWorkspaceFolder(vscode.Uri.parse(element))!;
+  }
+  return folder;
 }
 
 /**